manus_agents/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
feedparser
python-dotenv
aiofiles
pydantic
mysql-connector-python
Pillow
